Merge the fully validated consolidation and cross-platform hardening work.

+++ src/jsxer/nodes/ForInStatement.h
@@ -20,7 +20,7 @@ namespace jsxer::nodes {
         decoders::LineInfo bodyInfo;
         AstOpNode loopVariable;
         AstOpNode objExpression;
-        unsigned long length{};
+        size_t length{};
         string id;
         bool forEach = false;
     };
